requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pandas==2.2.0
scikit-learn==1.4.0
python-multipart==0.0.6
mlflow==2.11.1  # to load models from mlfow registry
setuptools==69.0.3
boto3==1.34.51  # mlflow host files on s3
python-gitlab==4.4.0  # to trigger pipeline from API
evidently==0.4.19  # to run data checks in CI